modify route of static folder
