+++ templates/attraction.html
@@ -5,7 +5,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title></title>
-    <link rel="stylesheet" href="/static/css/attraction.css">
+    <link rel="stylesheet" href="/css/reset.css">
+    <link rel="stylesheet" href="/css/base.css">
+    <link rel="stylesheet" href="/css/attraction.css">
     <!-- import Noto Sans TC -->
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -15,21 +17,13 @@
 </head>
 <body>
     <div class="wrapper">
-        <nav>
-            <div class="outer">
-                <a class="title" href="http://54.199.123.84:3000/">台北一日遊</a>
-                <ul class="menu">
-                    <li class="list1"><a href="#">預定行程</a></li>
-                    <li class="list2"><a href="#">登入/註冊</a></li>
-                </ul>
-            </div>
-        </nav>
+        <!-- Navbar : import from base.js -->
         
         <section>
             <div class="imgBox">
                 <div class="arrow">
-                    <img src="/static/image/btn_leftArrow.png" alt="">
-                    <img src="/static/image/btn_rightArrow.png" alt="">
+                    <img src="/image/btn_leftArrow.png" alt="">
+                    <img src="/image/btn_rightArrow.png" alt="">
                 </div>
             </div>
 
@@ -59,15 +53,16 @@
         <hr>
 
         <main>
-            <div class="headline address">景點地址：</div>
-            <div class="headline transport">交通方式：</div>
+            <div class="headline">景點地址：</div>
+            <div class="headline">交通方式：</div>
         </main>
         
-        <button class="topButton"><i class="fa-solid fa-angles-up"></i></button>
+        <!-- toTopButton : import from base.js -->
         
         <footer><p>COPYRIGHT © 2022 台北一日遊</p></footer>
     </div>
 
-    <script src="/static/js/attraction.js"></script>
+    <script src="/js/base.js"></script>
+    <script src="/js/attraction.js"></script>
 </body>
 </html>
