+++ templates/index.html
@@ -5,7 +5,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Taipei Trip 台北一日遊</title>
-    <link rel="stylesheet" href="/static/css/index.css">
+    <link rel="stylesheet" href="/css/reset.css">
+    <link rel="stylesheet" href="/css/index.css">
+    <link rel="stylesheet" href="/css/base.css">
     <!-- import Noto Sans TC -->
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -15,25 +17,17 @@
 </head>
 <body>
     <div class="wrapper">
-        <nav>
-            <div class="outer">
-                <a class="title" href="http://54.199.123.84:3000/">台北一日遊</a>
-                <ul class="menu">
-                    <li class="list1"><a href="#">預定行程</a></li>
-                    <li class="list2"><a href="#">登入/註冊</a></li>
-                </ul>
-            </div>
-        </nav>
+        <!-- Navbar : import from base.js -->
         
         <div class="searchContainer">
-            <img src="/static//image/welcome.png" class="backImg" alt="">
+            <img src="/image/welcome.png" class="backImg" alt="">
             <div class="searchBox">
                 <h1>輕鬆享受台北一日悠閒</h1>
                 <p>探索每個角落，體驗城市的深度旅遊行程</p>
                 <div class="searchBar">
                     <input type="text" class="searchInput" placeholder="輸入景點名稱查詢">
                     <button class="searchIcon">
-                        <img src="/static/image/icon_search.png" alt="">
+                        <img src="/image/icon_search.png" alt="">
                     </button>
                 </div>
                 <div class="categoryBox"></div>
@@ -42,11 +36,12 @@
 
         <main></main>
 
-        <button class="topButton"><i class="fa-solid fa-angles-up"></i></button>
+        <!-- toTopButton : import from base.js -->
         
         <footer><p>COPYRIGHT © 2022 台北一日遊</p></footer>
     </div>
 
-    <script src="/static/js/index.js"></script>
+    <script src="/js/base.js"></script>
+    <script src="/js/index.js"></script>
 </body>
 </html>
